Rename libopenmc.py capi.py
